fix(desktop): serialize heartbeat task execution
Problem:
- Concurrent desktop processes could execute the same due task, and every-minute cron tasks could be skipped after the first run.
- Concurrent frontend full-table mutations could lose updates, while the PR exceeded locale and repository capacity gates.

Root cause:
- The task lease ended before run state was durably persisted, due checks were not revalidated after lease handoff, and cron deduplication used only the previous run timestamp.
- Frontend mutations shared stale snapshots, and feature-only translations were bundled into every global locale chunk.

Fix:
- Hold the task lease through state persistence, re-read and revalidate after acquisition, and deduplicate cron runs by occurrence minute.
- Serialize frontend mutations against the latest authoritative task list and keep failed drafts visible.
- Split heartbeat owners and tests into focused files, keep feature translations in the lazy Heartbeat chunk, and restore the existing locale budgets.

Verification:
- cd desktop && go test ./... -count=1
- cd desktop && go test -race ./... -count=1
- pnpm test -- --keep-going (206 suites)
- pnpm build
- go run ./tools/repolint
- git diff --cached --check
