+++ desktop/frontend/src/custom/features/heartbeat/heartbeat.i18n.ts
@@ -1,0 +1,230 @@
+import { useCallback } from "react";
+import { useI18n, type DictKey, type Locale } from "../../../lib/i18n";
+
+export const heartbeatFeatureEn = {
+  "heartbeat.addTaskToScope": "Add task in {name}",
+  "heartbeat.btnNew": "New",
+  "heartbeat.btnPause": "Pause",
+  "heartbeat.btnStart": "Start",
+  "heartbeat.clickPause": "Click to pause",
+  "heartbeat.clickStart": "Click to start",
+  "heartbeat.cronAt": "at {time}",
+  "heartbeat.cronEveryHour": "every {n} hour(s)",
+  "heartbeat.cronEveryMin": "every {n} min",
+  "heartbeat.cronHourly": "hourly",
+  "heartbeat.cronNextRun": "Next run:",
+  "heartbeat.cronOnHour": "at minute {n} of every hour",
+  "heartbeat.cronPlaceholder": "0 * * * *",
+  "heartbeat.cronWeekdayFri": "Fri",
+  "heartbeat.cronWeekdayMon": "Mon",
+  "heartbeat.cronWeekdaySat": "Sat",
+  "heartbeat.cronWeekdaySun": "Sun",
+  "heartbeat.cronWeekdayThu": "Thu",
+  "heartbeat.cronWeekdayTue": "Tue",
+  "heartbeat.cronWeekdayWed": "Wed",
+  "heartbeat.cycleTime": "Time",
+  "heartbeat.cycleTimeAt": " (Time: {time})",
+  "heartbeat.daysAgo": "{n} d ago",
+  "heartbeat.detailTitle": "Task detail",
+  "heartbeat.discardChanges": "Discard unsaved changes?",
+  "heartbeat.dueSoon": "due soon",
+  "heartbeat.filterStatus": "Status",
+  "heartbeat.freqCron": "Custom",
+  "heartbeat.frequencyConversionFailed": "This schedule cannot be converted without changing when it runs.",
+  "heartbeat.heroSubtitle": "Let AI schedule tasks, set reminders, or monitor updates",
+  "heartbeat.heroTitle": "Automation",
+  "heartbeat.hoursAgo": "{n} hr ago",
+  "heartbeat.joinComma": ", ",
+  "heartbeat.later": "",
+  "heartbeat.minutesAgo": "{n} min ago",
+  "heartbeat.monthDay": "",
+  "heartbeat.newTask": "New",
+  "heartbeat.nextRun": "Next run in",
+  "heartbeat.remaining": "in ",
+  "heartbeat.runHistory": "Run history",
+  "heartbeat.runHistoryEmpty": "No runs yet",
+  "heartbeat.runHistoryNoTopic": "No conversation",
+  "heartbeat.saveFailed": "Could not save because automations changed elsewhere. Your draft is still here; review it and try again.",
+  "heartbeat.selectTask": "Select a task to view details",
+  "heartbeat.statusDisabled": "Paused",
+  "heartbeat.statusEnabled": "Active",
+  "heartbeat.sugDailyReview": "Daily review",
+  "heartbeat.sugDailyReviewDesc": "Read today's conversations across projects and review what was done, what's left, and tomorrow's plan",
+  "heartbeat.sugDailyReviewFreq": "Daily at 20:00",
+  "heartbeat.sugDailyReviewPrompt": "Read all of today's conversations and produce a multi-project review: what was completed, what issues remain, and what to do tomorrow. Output a structured review.",
+  "heartbeat.sugDownloads": "Downloads folder tidy-up report",
+  "heartbeat.sugDownloadsDesc": "Scan the downloads folder read-only and produce a categorized tidy-up report (no file moves)",
+  "heartbeat.sugDownloadsFreq": "Fridays at 16:00",
+  "heartbeat.sugDownloadsPrompt": "Scan the downloads folder of the current OS read-only (macOS/Linux: ~/Downloads, Windows: %USERPROFILE%\\Downloads), group files by type, and output a tidy-up suggestion report. Do not move or delete any files.",
+  "heartbeat.sugProductUpdate": "Product update digest",
+  "heartbeat.sugProductUpdateDesc": "Fetch the latest Reasonix release notes and learn about new capabilities",
+  "heartbeat.sugProductUpdateFreq": "Daily at 12:00",
+  "heartbeat.sugProductUpdatePrompt": "Visit https://reasonix.io/changelog/ to read the latest release notes, highlight recent new capabilities (New) and notable fixes in a bullet summary of at most 10 items. If the page is unreachable, fall back to release-notes/releases.json in the root of the local DeepSeek-Reasonix repository (if unsure where the repo is, locate the file with find).",
+  "heartbeat.suggestions": "Suggestions",
+  "heartbeat.suggestionsHint": "Create common automation tasks in one click",
+  "heartbeat.unitDay": "d",
+  "heartbeat.viewFlat": "Flat list",
+  "heartbeat.viewGrouped": "Group by project",
+} as const;
+
+export type HeartbeatFeatureKey = keyof typeof heartbeatFeatureEn;
+export type HeartbeatTranslationKey = DictKey | HeartbeatFeatureKey;
+export type HeartbeatTranslator = (
+  key: HeartbeatTranslationKey,
+  vars?: Record<string, string | number>,
+) => string;
+
+const heartbeatFeatureZh = {
+  "heartbeat.addTaskToScope": "在 {name} 中添加任务",
+  "heartbeat.btnNew": "新建",
+  "heartbeat.btnPause": "暂停",
+  "heartbeat.btnStart": "启动",
+  "heartbeat.clickPause": "点击暂停",
+  "heartbeat.clickStart": "点击启动",
+  "heartbeat.cronAt": "在 {time}",
+  "heartbeat.cronEveryHour": "每{n}小时",
+  "heartbeat.cronEveryMin": "每{n}分钟",
+  "heartbeat.cronHourly": "整点",
+  "heartbeat.cronNextRun": "下次运行:",
+  "heartbeat.cronOnHour": "每小时 {n} 分",
+  "heartbeat.cronPlaceholder": "0 * * * *",
+  "heartbeat.cronWeekdayFri": "周五",
+  "heartbeat.cronWeekdayMon": "周一",
+  "heartbeat.cronWeekdaySat": "周六",
+  "heartbeat.cronWeekdaySun": "周日",
+  "heartbeat.cronWeekdayThu": "周四",
+  "heartbeat.cronWeekdayTue": "周二",
+  "heartbeat.cronWeekdayWed": "周三",
+  "heartbeat.cycleTime": "时间",
+  "heartbeat.cycleTimeAt": "（时间：{time}）",
+  "heartbeat.daysAgo": "{n} 天前",
+  "heartbeat.detailTitle": "任务详情",
+  "heartbeat.discardChanges": "放弃未保存的修改？",
+  "heartbeat.dueSoon": "即将触发",
+  "heartbeat.filterStatus": "状态筛选",
+  "heartbeat.freqCron": "自定义",
+  "heartbeat.frequencyConversionFailed": "该调度无法在不改变运行时间的前提下转换。",
+  "heartbeat.heroSubtitle": "让 AI 安排任务、设置提醒或监测更新",
+  "heartbeat.heroTitle": "自动化",
+  "heartbeat.hoursAgo": "{n} 小时前",
+  "heartbeat.joinComma": "、",
+  "heartbeat.later": "后",
+  "heartbeat.minutesAgo": "{n} 分钟前",
+  "heartbeat.monthDay": "日",
+  "heartbeat.newTask": "新建",
+  "heartbeat.nextRun": "下次运行",
+  "heartbeat.remaining": "剩余",
+  "heartbeat.runHistory": "运行历史记录",
+  "heartbeat.runHistoryEmpty": "暂无执行记录",
+  "heartbeat.runHistoryNoTopic": "无会话",
+  "heartbeat.saveFailed": "自动化配置已在其他位置更改，本次保存失败。草稿仍已保留，请确认后重试。",
+  "heartbeat.selectTask": "选择一个任务查看详情",
+  "heartbeat.statusDisabled": "已暂停",
+  "heartbeat.statusEnabled": "活跃",
+  "heartbeat.sugDailyReview": "每日复盘",
+  "heartbeat.sugDailyReviewDesc": "读取今天所有项目的对话，复盘当天完成的工作、遗留问题与明日计划",
+  "heartbeat.sugDailyReviewFreq": "每天 20:00",
+  "heartbeat.sugDailyReviewPrompt": "读取今天的全部对话，按项目进行多项目复盘：今天完成了什么、有什么遗留问题、明天计划做什么。输出结构化复盘。",
+  "heartbeat.sugDownloads": "下载文件夹整理报告",
+  "heartbeat.sugDownloadsDesc": "只读扫描下载文件夹，生成分类整理建议（不移动文件）",
+  "heartbeat.sugDownloadsFreq": "每周五 16:00",
+  "heartbeat.sugDownloadsPrompt": "只读扫描当前操作系统的下载文件夹（macOS/Linux 为 ~/Downloads，Windows 为 %USERPROFILE%\\Downloads），按类型整理出分类清单，输出整理建议报告。不要移动或删除任何文件。",
+  "heartbeat.sugProductUpdate": "产品更新速览",
+  "heartbeat.sugProductUpdateDesc": "查询 Reasonix 最新发布说明，了解新能力与用法",
+  "heartbeat.sugProductUpdateFreq": "每天 12:00",
+  "heartbeat.sugProductUpdatePrompt": "访问 https://reasonix.io/changelog/ 查看最新的发布说明，重点介绍最近版本的新能力（New/新增）和重要修复，用要点式摘要输出，控制在 10 条以内。如果页面无法访问，改用本地 DeepSeek-Reasonix 代码仓库根目录下的 release-notes/releases.json 兜底（若不确定仓库位置，用 find 命令搜索该文件）。",
+  "heartbeat.suggestions": "建议",
+  "heartbeat.suggestionsHint": "一键创建常用自动化任务",
+  "heartbeat.unitDay": "天",
+  "heartbeat.viewFlat": "纯列表",
+  "heartbeat.viewGrouped": "按项目分组",
+} satisfies Record<HeartbeatFeatureKey, string>;
+
+const heartbeatFeatureZhTW = {
+  "heartbeat.addTaskToScope": "在 {name} 中添加任务",
+  "heartbeat.btnNew": "新增",
+  "heartbeat.btnPause": "暫停",
+  "heartbeat.btnStart": "啟動",
+  "heartbeat.clickPause": "點擊暫停",
+  "heartbeat.clickStart": "點擊啟動",
+  "heartbeat.cronAt": "在 {time}",
+  "heartbeat.cronEveryHour": "每{n}小時",
+  "heartbeat.cronEveryMin": "每{n}分鐘",
+  "heartbeat.cronHourly": "整點",
+  "heartbeat.cronNextRun": "下次運行:",
+  "heartbeat.cronOnHour": "每小時 {n} 分",
+  "heartbeat.cronPlaceholder": "0 * * * *",
+  "heartbeat.cronWeekdayFri": "週五",
+  "heartbeat.cronWeekdayMon": "週一",
+  "heartbeat.cronWeekdaySat": "週六",
+  "heartbeat.cronWeekdaySun": "週日",
+  "heartbeat.cronWeekdayThu": "週四",
+  "heartbeat.cronWeekdayTue": "週二",
+  "heartbeat.cronWeekdayWed": "週三",
+  "heartbeat.cycleTime": "時間",
+  "heartbeat.cycleTimeAt": "（時間：{time}）",
+  "heartbeat.daysAgo": "{n} 天前",
+  "heartbeat.detailTitle": "任務詳情",
+  "heartbeat.discardChanges": "放棄未儲存的修改？",
+  "heartbeat.dueSoon": "即將觸發",
+  "heartbeat.filterStatus": "狀態篩選",
+  "heartbeat.freqCron": "自訂",
+  "heartbeat.frequencyConversionFailed": "該排程無法在不改變執行時間的前提下轉換。",
+  "heartbeat.heroSubtitle": "讓 AI 安排任務、設定提醒或監測更新",
+  "heartbeat.heroTitle": "自動化",
+  "heartbeat.hoursAgo": "{n} 小時前",
+  "heartbeat.joinComma": "、",
+  "heartbeat.later": "後",
+  "heartbeat.minutesAgo": "{n} 分鐘前",
+  "heartbeat.monthDay": "日",
+  "heartbeat.newTask": "新建",
+  "heartbeat.nextRun": "下次運行",
+  "heartbeat.remaining": "剩餘",
+  "heartbeat.runHistory": "運行歷史記錄",
+  "heartbeat.runHistoryEmpty": "暫無執行記錄",
+  "heartbeat.runHistoryNoTopic": "無會話",
+  "heartbeat.saveFailed": "自動化設定已在其他位置變更，本次儲存失敗。草稿仍已保留，請確認後重試。",
+  "heartbeat.selectTask": "選擇一個任務查看詳情",
+  "heartbeat.statusDisabled": "已暫停",
+  "heartbeat.statusEnabled": "活躍",
+  "heartbeat.sugDailyReview": "每日覆盤",
+  "heartbeat.sugDailyReviewDesc": "讀取今天所有專案的對話，覆盤當天完成的工作、遺留問題與明日計劃",
+  "heartbeat.sugDailyReviewFreq": "每天 20:00",
+  "heartbeat.sugDailyReviewPrompt": "讀取今天的全部對話，按專案進行多專案覆盤：今天完成了什麼、有什麼遺留問題、明天計劃做什麼。輸出結構化覆盤。",
+  "heartbeat.sugDownloads": "下載資料夾整理報告",
+  "heartbeat.sugDownloadsDesc": "唯讀掃描下載資料夾，產生分類整理建議（不移動檔案）",
+  "heartbeat.sugDownloadsFreq": "每週五 16:00",
+  "heartbeat.sugDownloadsPrompt": "唯讀掃描目前作業系統的下載資料夾（macOS/Linux 為 ~/Downloads，Windows 為 %USERPROFILE%\\Downloads），按類型整理出分類清單，輸出整理建議報告。不要移動或刪除任何檔案。",
+  "heartbeat.sugProductUpdate": "產品更新速覽",
+  "heartbeat.sugProductUpdateDesc": "查詢 Reasonix 最新發佈說明，了解新能力與用法",
+  "heartbeat.sugProductUpdateFreq": "每天 12:00",
+  "heartbeat.sugProductUpdatePrompt": "訪問 https://reasonix.io/changelog/ 查看最新的發佈說明，重點介紹最近版本的新能力（New/新增）和重要修復，用要點式摘要輸出，控制在 10 條以內。如果頁面無法訪問，改用本地 DeepSeek-Reasonix 程式碼倉庫根目錄下的 release-notes/releases.json 兜底（若不確定倉庫位置，用 find 指令搜尋該檔案）。",
+  "heartbeat.suggestions": "建議",
+  "heartbeat.suggestionsHint": "一鍵建立常用自動化任務",
+  "heartbeat.unitDay": "天",
+  "heartbeat.viewFlat": "純列表",
+  "heartbeat.viewGrouped": "按專案分組",
+} satisfies Record<HeartbeatFeatureKey, string>;
+
+export const heartbeatFeatureKeys = Object.keys(heartbeatFeatureEn) as HeartbeatFeatureKey[];
+
+const heartbeatFeatureDictionaries: Record<Locale, Record<HeartbeatFeatureKey, string>> = {
+  en: heartbeatFeatureEn,
+  zh: heartbeatFeatureZh,
+  "zh-TW": heartbeatFeatureZhTW,
+};
+
+function interpolate(message: string, vars?: Record<string, string | number>): string {
+  if (!vars) return message;
+  return message.replace(/\{(\w+)\}/g, (_, key) => (
+    vars[key] === undefined ? `{${key}}` : String(vars[key])
+  ));
+}
+
+export function useHeartbeatT(): HeartbeatTranslator {
+  const { locale, t } = useI18n();
+  return useCallback<HeartbeatTranslator>((key, vars) => {
+    const message = heartbeatFeatureDictionaries[locale][key as HeartbeatFeatureKey];
+    return message === undefined ? t(key as DictKey, vars) : interpolate(message, vars);
+  }, [locale, t]);
+}
