+++ desktop/frontend/src/locales/en.ts
@@ -907,12 +907,6 @@ export const en = {
   "heartbeat.title": "Heartbeat Tasks",
   "heartbeat.scheduler": "Scheduled Tasks",
   "heartbeat.addTask": "Add task",
-  "heartbeat.btnNew": "New",
-  "heartbeat.viewFlat": "Flat list",
-  "heartbeat.viewGrouped": "Group by project",
-  "heartbeat.heroTitle": "Automation",
-  "heartbeat.heroSubtitle": "Let AI schedule tasks, set reminders, or monitor updates",
-  "heartbeat.detailTitle": "Task detail",
   "heartbeat.edit": "Edit",
   "heartbeat.delete": "Delete",
   "heartbeat.runNow": "Run now",
@@ -926,19 +920,10 @@ export const en = {
   "heartbeat.promptPlaceholder": "e.g. Check for any new issues on GitHub and summarize them.",
   "heartbeat.fieldInterval": "Frequency",
   "heartbeat.enabled": "Enabled",
-  "heartbeat.statusEnabled": "Active",
-  "heartbeat.statusDisabled": "Paused",
-  "heartbeat.newTask": "New",
-  "heartbeat.clickPause": "Click to pause",
-  "heartbeat.clickStart": "Click to start",
-  "heartbeat.btnPause": "Pause",
-  "heartbeat.btnStart": "Start",
   "heartbeat.status": "Status",
   "heartbeat.optional": "Optional",
   "heartbeat.disable": "Disable",
   "heartbeat.save": "Save",
-  "heartbeat.saveFailed": "Could not save because automations changed elsewhere. Your draft is still here; review it and try again.",
-  "heartbeat.frequencyConversionFailed": "This schedule cannot be converted without changing when it runs.",
   "heartbeat.add": "Add",
   "heartbeat.neverRun": "Never run",
   "heartbeat.untitled": "(untitled)",
@@ -955,50 +940,23 @@ export const en = {
   "heartbeat.currentWorkspace": "current",
   "heartbeat.freqCycle": "Cycle",
   "heartbeat.freqInterval": "Interval",
-  "heartbeat.freqCron": "Custom",
-  "heartbeat.cronPlaceholder": "0 * * * *",
-  "heartbeat.cronEveryMin": "every {n} min",
-  "heartbeat.cronEveryHour": "every {n} hour(s)",
-  "heartbeat.cronAt": "at {time}",
-  "heartbeat.cronHourly": "hourly",
-  "heartbeat.cronOnHour": "at minute {n} of every hour",
-  "heartbeat.cronWeekdaySun": "Sun",
-  "heartbeat.cronWeekdayMon": "Mon",
-  "heartbeat.cronWeekdayTue": "Tue",
-  "heartbeat.cronWeekdayWed": "Wed",
-  "heartbeat.cronWeekdayThu": "Thu",
-  "heartbeat.cronWeekdayFri": "Fri",
-  "heartbeat.cronWeekdaySat": "Sat",
-  "heartbeat.joinComma": ", ",
-  "heartbeat.cronNextRun": "Next run:",
-    "heartbeat.freqDaily": "Every day",
+  "heartbeat.freqDaily": "Every day",
   "heartbeat.freqEvery": "Every",
   "heartbeat.everyJoiner": " ",
   "heartbeat.unitSec": "sec",
   "heartbeat.unitMin": "min",
   "heartbeat.unitHour": "h",
-  "heartbeat.unitDay": "d",
-  "heartbeat.remaining": "in ",
-  "heartbeat.nextRun": "Next run in",
-  "heartbeat.later": "",
-  "heartbeat.dueSoon": "due soon",
   "heartbeat.timeWindow": "Time window",
   "heartbeat.clearTimeWindow": "Clear time window",
   "heartbeat.cycleDaily": "Daily",
-  "heartbeat.cycleTime": "Time",
   "heartbeat.cycleWeekly": "Weekly",
   "heartbeat.cycleBiweekly": "Bi-weekly",
   "heartbeat.cycleMonthly": "Monthly",
   "heartbeat.cycleYearly": "Yearly",
-  "heartbeat.monthDay": "",
-  "heartbeat.cycleTimeAt": " (Time: {time})",
   "heartbeat.justNow": "just now",
   "heartbeat.due": "due soon",
   "heartbeat.soon": "in 1 min",
-                "heartbeat.removeTimeWindow": "Remove time window",
-  "heartbeat.selectTask": "Select a task to view details",
-  "heartbeat.discardChanges": "Discard unsaved changes?",
-  "heartbeat.addTaskToScope": "Add task in {name}",
+  "heartbeat.removeTimeWindow": "Remove time window",
   "heartbeat.minLater": "m left",
   "heartbeat.hourLater": "h left",
   "heartbeat.minAgo": "m ago",
@@ -1007,30 +965,9 @@ export const en = {
   "heartbeat.filterAll": "All",
   "heartbeat.filterEnabled": "Enabled",
   "heartbeat.filterDisabled": "Paused",
-  "heartbeat.filterStatus": "Status",
   "heartbeat.searchPlaceholder": "Search tasks…",
   "heartbeat.filterAllProjects": "All projects",
   "heartbeat.noMatchingTasks": "No matching tasks",
-  "heartbeat.suggestions": "Suggestions",
-  "heartbeat.suggestionsHint": "Create common automation tasks in one click",
-  "heartbeat.runHistory": "Run history",
-  "heartbeat.runHistoryEmpty": "No runs yet",
-  "heartbeat.runHistoryNoTopic": "No conversation",
-  "heartbeat.minutesAgo": "{n} min ago",
-  "heartbeat.hoursAgo": "{n} hr ago",
-  "heartbeat.daysAgo": "{n} d ago",
-  "heartbeat.sugDailyReview": "Daily review",
-  "heartbeat.sugDailyReviewDesc": "Read today's conversations across projects and review what was done, what's left, and tomorrow's plan",
-  "heartbeat.sugDailyReviewFreq": "Daily at 20:00",
-  "heartbeat.sugDailyReviewPrompt": "Read all of today's conversations and produce a multi-project review: what was completed, what issues remain, and what to do tomorrow. Output a structured review.",
-  "heartbeat.sugProductUpdate": "Product update digest",
-  "heartbeat.sugProductUpdateDesc": "Fetch the latest Reasonix release notes and learn about new capabilities",
-  "heartbeat.sugProductUpdateFreq": "Daily at 12:00",
-  "heartbeat.sugProductUpdatePrompt": "Visit https://reasonix.io/changelog/ to read the latest release notes, highlight recent new capabilities (New) and notable fixes in a bullet summary of at most 10 items. If the page is unreachable, fall back to release-notes/releases.json in the root of the local DeepSeek-Reasonix repository (if unsure where the repo is, locate the file with find).",
-  "heartbeat.sugDownloads": "Downloads folder tidy-up report",
-  "heartbeat.sugDownloadsDesc": "Scan the downloads folder read-only and produce a categorized tidy-up report (no file moves)",
-  "heartbeat.sugDownloadsFreq": "Fridays at 16:00",
-  "heartbeat.sugDownloadsPrompt": "Scan the downloads folder of the current OS read-only (macOS/Linux: ~/Downloads, Windows: %USERPROFILE%\\Downloads), group files by type, and output a tidy-up suggestion report. Do not move or delete any files.",
   "heartbeat.weekdayMon": "Mon",
   "heartbeat.weekdayTue": "Tue",
   "heartbeat.weekdayWed": "Wed",
