+++ desktop/frontend/src/locales/zh-TW.ts
@@ -677,12 +677,6 @@ export const zhTW: Record<DictKey, string> = {
   "heartbeat.title": "心跳任務",
   "heartbeat.scheduler": "定時任務",
   "heartbeat.addTask": "添加任務",
-  "heartbeat.btnNew": "新增",
-  "heartbeat.viewFlat": "純列表",
-  "heartbeat.viewGrouped": "按專案分組",
-  "heartbeat.heroTitle": "自動化",
-  "heartbeat.heroSubtitle": "讓 AI 安排任務、設定提醒或監測更新",
-  "heartbeat.detailTitle": "任務詳情",
   "heartbeat.edit": "編輯",
   "heartbeat.delete": "刪除",
   "heartbeat.runNow": "立即執行",
@@ -696,19 +690,10 @@ export const zhTW: Record<DictKey, string> = {
   "heartbeat.promptPlaceholder": "例如：檢查 GitHub 上是否有新的 issue 並總結。",
   "heartbeat.fieldInterval": "頻率",
   "heartbeat.enabled": "啟用",
-  "heartbeat.statusEnabled": "活躍",
-  "heartbeat.statusDisabled": "已暫停",
-  "heartbeat.newTask": "新建",
-  "heartbeat.clickPause": "點擊暫停",
-  "heartbeat.clickStart": "點擊啟動",
-  "heartbeat.btnPause": "暫停",
-  "heartbeat.btnStart": "啟動",
   "heartbeat.status": "狀態",
   "heartbeat.optional": "可選",
   "heartbeat.disable": "停用",
   "heartbeat.save": "保存",
-  "heartbeat.saveFailed": "自動化設定已在其他位置變更，本次儲存失敗。草稿仍已保留，請確認後重試。",
-  "heartbeat.frequencyConversionFailed": "該排程無法在不改變執行時間的前提下轉換。",
   "heartbeat.add": "添加",
   "heartbeat.neverRun": "從未執行",
   "heartbeat.untitled": "（無標題）",
@@ -725,50 +710,23 @@ export const zhTW: Record<DictKey, string> = {
   "heartbeat.currentWorkspace": "當前",
   "heartbeat.freqCycle": "週期",
   "heartbeat.freqInterval": "間隔",
-  "heartbeat.freqCron": "自訂",
-  "heartbeat.cronPlaceholder": "0 * * * *",
-  "heartbeat.cronEveryMin": "每{n}分鐘",
-  "heartbeat.cronEveryHour": "每{n}小時",
-  "heartbeat.cronAt": "在 {time}",
-  "heartbeat.cronHourly": "整點",
-  "heartbeat.cronOnHour": "每小時 {n} 分",
-  "heartbeat.cronWeekdaySun": "週日",
-  "heartbeat.cronWeekdayMon": "週一",
-  "heartbeat.cronWeekdayTue": "週二",
-  "heartbeat.cronWeekdayWed": "週三",
-  "heartbeat.cronWeekdayThu": "週四",
-  "heartbeat.cronWeekdayFri": "週五",
-  "heartbeat.cronWeekdaySat": "週六",
-  "heartbeat.joinComma": "、",
-  "heartbeat.cronNextRun": "下次運行:",
-    "heartbeat.freqDaily": "每天",
+  "heartbeat.freqDaily": "每天",
   "heartbeat.freqEvery": "每",
   "heartbeat.everyJoiner": "",
   "heartbeat.unitSec": "秒",
   "heartbeat.unitMin": "分鐘",
   "heartbeat.unitHour": "小時",
-  "heartbeat.unitDay": "天",
-  "heartbeat.remaining": "剩餘",
-  "heartbeat.nextRun": "下次運行",
-  "heartbeat.later": "後",
-  "heartbeat.dueSoon": "即將觸發",
   "heartbeat.timeWindow": "時間窗口",
   "heartbeat.clearTimeWindow": "清除時間窗口",
   "heartbeat.cycleDaily": "每天",
-  "heartbeat.cycleTime": "時間",
   "heartbeat.cycleWeekly": "每週",
   "heartbeat.cycleBiweekly": "雙週",
   "heartbeat.cycleMonthly": "每月",
   "heartbeat.cycleYearly": "每年",
-  "heartbeat.monthDay": "日",
-  "heartbeat.cycleTimeAt": "（時間：{time}）",
   "heartbeat.justNow": "剛剛",
   "heartbeat.due": "即將執行",
   "heartbeat.soon": "1分鐘內",
-                "heartbeat.removeTimeWindow": "移除時間區間",
-  "heartbeat.selectTask": "選擇一個任務查看詳情",
-  "heartbeat.discardChanges": "放棄未儲存的修改？",
-  "heartbeat.addTaskToScope": "在 {name} 中添加任務",
+  "heartbeat.removeTimeWindow": "移除時間區間",
   "heartbeat.minLater": "分鐘後",
   "heartbeat.hourLater": "小時後",
   "heartbeat.minAgo": "分鐘前",
@@ -777,30 +735,9 @@ export const zhTW: Record<DictKey, string> = {
   "heartbeat.filterAll": "全部",
   "heartbeat.filterEnabled": "已開啟",
   "heartbeat.filterDisabled": "已暫停",
-  "heartbeat.filterStatus": "狀態篩選",
   "heartbeat.searchPlaceholder": "搜索任務名稱…",
   "heartbeat.filterAllProjects": "全部項目",
   "heartbeat.noMatchingTasks": "沒有符合的任務",
-  "heartbeat.suggestions": "建議",
-  "heartbeat.suggestionsHint": "一鍵建立常用自動化任務",
-  "heartbeat.runHistory": "運行歷史記錄",
-  "heartbeat.runHistoryEmpty": "暫無執行記錄",
-  "heartbeat.runHistoryNoTopic": "無會話",
-  "heartbeat.minutesAgo": "{n} 分鐘前",
-  "heartbeat.hoursAgo": "{n} 小時前",
-  "heartbeat.daysAgo": "{n} 天前",
-  "heartbeat.sugDailyReview": "每日覆盤",
-  "heartbeat.sugDailyReviewDesc": "讀取今天所有專案的對話，覆盤當天完成的工作、遺留問題與明日計劃",
-  "heartbeat.sugDailyReviewFreq": "每天 20:00",
-  "heartbeat.sugDailyReviewPrompt": "讀取今天的全部對話，按專案進行多專案覆盤：今天完成了什麼、有什麼遺留問題、明天計劃做什麼。輸出結構化覆盤。",
-  "heartbeat.sugProductUpdate": "產品更新速覽",
-  "heartbeat.sugProductUpdateDesc": "查詢 Reasonix 最新發佈說明，了解新能力與用法",
-  "heartbeat.sugProductUpdateFreq": "每天 12:00",
-  "heartbeat.sugProductUpdatePrompt": "訪問 https://reasonix.io/changelog/ 查看最新的發佈說明，重點介紹最近版本的新能力（New/新增）和重要修復，用要點式摘要輸出，控制在 10 條以內。如果頁面無法訪問，改用本地 DeepSeek-Reasonix 程式碼倉庫根目錄下的 release-notes/releases.json 兜底（若不確定倉庫位置，用 find 指令搜尋該檔案）。",
-  "heartbeat.sugDownloads": "下載資料夾整理報告",
-  "heartbeat.sugDownloadsDesc": "唯讀掃描下載資料夾，產生分類整理建議（不移動檔案）",
-  "heartbeat.sugDownloadsFreq": "每週五 16:00",
-  "heartbeat.sugDownloadsPrompt": "唯讀掃描目前作業系統的下載資料夾（macOS/Linux 為 ~/Downloads，Windows 為 %USERPROFILE%\\Downloads），按類型整理出分類清單，輸出整理建議報告。不要移動或刪除任何檔案。",
   "heartbeat.weekdayMon": "週一",
   "heartbeat.weekdayTue": "週二",
   "heartbeat.weekdayWed": "週三",
