+++ desktop/frontend/src/locales/zh.ts
@@ -908,12 +908,6 @@ export const zh: Record<DictKey, string> = {
   "heartbeat.title": "心跳任务",
   "heartbeat.scheduler": "定时任务",
   "heartbeat.addTask": "添加任务",
-  "heartbeat.btnNew": "新建",
-  "heartbeat.viewFlat": "纯列表",
-  "heartbeat.viewGrouped": "按项目分组",
-  "heartbeat.heroTitle": "自动化",
-  "heartbeat.heroSubtitle": "让 AI 安排任务、设置提醒或监测更新",
-  "heartbeat.detailTitle": "任务详情",
   "heartbeat.edit": "编辑",
   "heartbeat.delete": "删除",
   "heartbeat.runNow": "立即执行",
@@ -927,19 +921,10 @@ export const zh: Record<DictKey, string> = {
   "heartbeat.promptPlaceholder": "例如：检查 GitHub 上是否有新的 issue 并总结。",
   "heartbeat.fieldInterval": "频率",
   "heartbeat.enabled": "启用",
-  "heartbeat.statusEnabled": "活跃",
-  "heartbeat.statusDisabled": "已暂停",
-  "heartbeat.newTask": "新建",
-  "heartbeat.clickPause": "点击暂停",
-  "heartbeat.clickStart": "点击启动",
-  "heartbeat.btnPause": "暂停",
-  "heartbeat.btnStart": "启动",
   "heartbeat.status": "状态",
   "heartbeat.optional": "可选",
   "heartbeat.disable": "禁用",
   "heartbeat.save": "保存",
-  "heartbeat.saveFailed": "自动化配置已在其他位置更改，本次保存失败。草稿仍已保留，请确认后重试。",
-  "heartbeat.frequencyConversionFailed": "该调度无法在不改变运行时间的前提下转换。",
   "heartbeat.add": "添加",
   "heartbeat.neverRun": "从未执行",
   "heartbeat.untitled": "（无标题）",
@@ -956,50 +941,23 @@ export const zh: Record<DictKey, string> = {
   "heartbeat.currentWorkspace": "当前",
   "heartbeat.freqCycle": "周期",
   "heartbeat.freqInterval": "间隔",
-  "heartbeat.freqCron": "自定义",
-  "heartbeat.cronPlaceholder": "0 * * * *",
-  "heartbeat.cronEveryMin": "每{n}分钟",
-  "heartbeat.cronEveryHour": "每{n}小时",
-  "heartbeat.cronAt": "在 {time}",
-  "heartbeat.cronHourly": "整点",
-  "heartbeat.cronOnHour": "每小时 {n} 分",
-  "heartbeat.cronWeekdaySun": "周日",
-  "heartbeat.cronWeekdayMon": "周一",
-  "heartbeat.cronWeekdayTue": "周二",
-  "heartbeat.cronWeekdayWed": "周三",
-  "heartbeat.cronWeekdayThu": "周四",
-  "heartbeat.cronWeekdayFri": "周五",
-  "heartbeat.cronWeekdaySat": "周六",
-  "heartbeat.joinComma": "、",
-  "heartbeat.cronNextRun": "下次运行:",
-    "heartbeat.freqDaily": "每天",
+  "heartbeat.freqDaily": "每天",
   "heartbeat.freqEvery": "每",
   "heartbeat.everyJoiner": "",
   "heartbeat.unitSec": "秒",
   "heartbeat.unitMin": "分钟",
   "heartbeat.unitHour": "小时",
-  "heartbeat.unitDay": "天",
-  "heartbeat.remaining": "剩余",
-  "heartbeat.nextRun": "下次运行",
-  "heartbeat.later": "后",
-  "heartbeat.dueSoon": "即将触发",
   "heartbeat.timeWindow": "时间窗口",
   "heartbeat.clearTimeWindow": "清除时间窗口",
   "heartbeat.cycleDaily": "每天",
-  "heartbeat.cycleTime": "时间",
   "heartbeat.cycleWeekly": "每周",
   "heartbeat.cycleBiweekly": "双周",
   "heartbeat.cycleMonthly": "每月",
   "heartbeat.cycleYearly": "每年",
-  "heartbeat.monthDay": "日",
-  "heartbeat.cycleTimeAt": "（时间：{time}）",
   "heartbeat.justNow": "刚刚",
   "heartbeat.due": "即将执行",
   "heartbeat.soon": "1分钟内",
-                "heartbeat.removeTimeWindow": "移除时间区间",
-  "heartbeat.selectTask": "选择一个任务查看详情",
-  "heartbeat.discardChanges": "放弃未保存的修改？",
-  "heartbeat.addTaskToScope": "在 {name} 中添加任务",
+  "heartbeat.removeTimeWindow": "移除时间区间",
   "heartbeat.minLater": "分钟后",
   "heartbeat.hourLater": "小时后",
   "heartbeat.minAgo": "分钟前",
@@ -1008,29 +966,8 @@ export const zh: Record<DictKey, string> = {
   "heartbeat.filterAll": "全部",
   "heartbeat.filterEnabled": "已开启",
   "heartbeat.filterDisabled": "已暂停",
-  "heartbeat.filterStatus": "状态筛选",
   "heartbeat.filterAllProjects": "全部项目",
   "heartbeat.noMatchingTasks": "没有匹配的任务",
-  "heartbeat.suggestions": "建议",
-  "heartbeat.suggestionsHint": "一键创建常用自动化任务",
-  "heartbeat.runHistory": "运行历史记录",
-  "heartbeat.runHistoryEmpty": "暂无执行记录",
-  "heartbeat.runHistoryNoTopic": "无会话",
-  "heartbeat.minutesAgo": "{n} 分钟前",
-  "heartbeat.hoursAgo": "{n} 小时前",
-  "heartbeat.daysAgo": "{n} 天前",
-  "heartbeat.sugDailyReview": "每日复盘",
-  "heartbeat.sugDailyReviewDesc": "读取今天所有项目的对话，复盘当天完成的工作、遗留问题与明日计划",
-  "heartbeat.sugDailyReviewFreq": "每天 20:00",
-  "heartbeat.sugDailyReviewPrompt": "读取今天的全部对话，按项目进行多项目复盘：今天完成了什么、有什么遗留问题、明天计划做什么。输出结构化复盘。",
-  "heartbeat.sugProductUpdate": "产品更新速览",
-  "heartbeat.sugProductUpdateDesc": "查询 Reasonix 最新发布说明，了解新能力与用法",
-  "heartbeat.sugProductUpdateFreq": "每天 12:00",
-  "heartbeat.sugProductUpdatePrompt": "访问 https://reasonix.io/changelog/ 查看最新的发布说明，重点介绍最近版本的新能力（New/新增）和重要修复，用要点式摘要输出，控制在 10 条以内。如果页面无法访问，改用本地 DeepSeek-Reasonix 代码仓库根目录下的 release-notes/releases.json 兜底（若不确定仓库位置，用 find 命令搜索该文件）。",
-  "heartbeat.sugDownloads": "下载文件夹整理报告",
-  "heartbeat.sugDownloadsDesc": "只读扫描下载文件夹，生成分类整理建议（不移动文件）",
-  "heartbeat.sugDownloadsFreq": "每周五 16:00",
-  "heartbeat.sugDownloadsPrompt": "只读扫描当前操作系统的下载文件夹（macOS/Linux 为 ~/Downloads，Windows 为 %USERPROFILE%\\Downloads），按类型整理出分类清单，输出整理建议报告。不要移动或删除任何文件。",
   "heartbeat.weekdayMon": "周一",
   "heartbeat.weekdayTue": "周二",
   "heartbeat.weekdayWed": "周三",
