+++ tools/repolint/baseline.json
@@ -67,10 +67,10 @@
       "essay": 2
     },
     "desktop/frontend/src/App.tsx": {
-      "file-size": 4683
+      "file-size": 4718
     },
     "desktop/frontend/src/__tests__/app-chrome-tabs.test.ts": {
-      "test-file-size": 19
+      "test-file-size": 27
     },
     "desktop/frontend/src/__tests__/capabilities-panel-actions.test.ts": {
       "test-file-size": 307
@@ -148,10 +148,10 @@
       "file-size": 1406
     },
     "desktop/frontend/src/custom/features/heartbeat/HeartbeatPanel.tsx": {
-      "file-size": 415
+      "file-size": 180
     },
     "desktop/frontend/src/lib/bridge.ts": {
-      "file-size": 4817
+      "file-size": 4824
     },
     "desktop/frontend/src/lib/crash.ts": {
       "file-size": 225
@@ -166,9 +166,7 @@
       "file-size": 4035
     },
     "desktop/heartbeat.go": {
-      "essay": 12,
-      "file-size": 349,
-      "function-size": 23
+      "essay": 7
     },
     "desktop/history_slice.go": {
       "banner": 2,
